auth_service/pyproject.toml
[tool.poetry]
name = "pet-microservice"
version = "0.1.0"
description = ""
authors = ["Nikita Duzhnov <duzhnov.nikita55@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
uvicorn = "^0.30.6"
gunicorn = "^23.0.0"
alembic = "^1.13.3"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
email-validator = "^2.0"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
